made about page editable
