+++ src/constants.js
@@ -1,4 +1,4 @@
 export default {
     year: '2019-2020 2학기 | 2nd Semester',
-    classes: ['유치가', '유치나', '기초1', '기초2', '기초3', '초급', '중급', '고급'],
+    classes: ['유치가 (금|토)', '유치나', '기초1 (금)', '기초1 (토)', '기초2 (금|토)', '기초3', '초급 (금|토)', '중급', '고급'],
 }
