+++ src/pages/Home.js
@@ -7,7 +7,10 @@ const Home = () => {
   return (
     <div>
         <br/>
-        <Alert variant='warning'>Closed</Alert>
+        <Alert variant='warning'>&#9432; The school is temporarily closed due to COVID-19.</Alert>
+        <Alert variant='success'>You may still access online classes by clicking 
+        <a href="/class"> here</a> and selecting your child's class. <br/><br/>
+        해당 자녀 <a href="/class">수업반</a>을 들어가시면 온라인 수업/숙제를 보실 수 있습니다. </Alert>
         <div className="jumbotron jumbotron-fluid bg-dark text-center">
         <div className="jumbotron-background">
             <img src="../students.jpg" className="blur img-fluid "/>
